requirements.txt
absl-py==2.2.1
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiokafka==0.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==5.0.1
atomicwrites==1.4.1
attrs==25.1.0
Automat==24.8.1
azure-core==1.32.0
azure-storage-blob==12.24.1
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
colorama==0.4.6
confluent-kafka==2.8.0
constantly==23.10.4
contourpy==1.3.1
Crawl4AI==0.6.3
cryptography==44.0.1
cssselect==1.2.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
elastic-transport==8.17.0
elasticsearch==8.17.1
elasticsearch-dsl==8.17.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fake-http-header==0.3.5
fake-useragent==2.0.3
fastapi==0.115.12
fastjsonschema==2.21.1
faust-cchardet==2.1.19
filelock==3.17.0
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.3.1
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.1
git-filter-repo==2.47.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.88.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.2.0
google-geo-type==0.3.12
google-maps-places==0.1.25
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
googlemaps==4.10.0
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
haversine==2.9.0
html2text==2024.2.26
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.29.1
humanize==4.12.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
iniconfig==2.1.0
instagram-private-api==1.6.0.0
ipython==8.12.3
isodate==0.7.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kafka==1.3.5
kafka-python==2.1.5
kafka-scrapy-connect==2.5.0
kiwisolver==1.4.8
litellm==1.61.13
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openai==1.63.0
openpyxl==3.1.5
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.4
pathlib_mate==1.3.2
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.8
playwright==1.50.0
pluggy==1.5.0
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.0
Protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
pyee==12.1.1
Pygments==2.19.1
PyMuPDF==1.25.3
pymupdf4llm==0.0.17
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyparsing==3.2.2
pypdfium2==4.30.1
pyperclip==1.9.0
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ulid==3.0.0
pytorch-crf==0.7.2
pytrends==4.9.2
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.4.0
queuelib==1.7.0
rank-bm25==0.2.2
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.12.0
scrapy-playwright==0.0.43
selenium==4.31.0
serpapi==0.1.5
service-identity==24.2.0
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.54
sqlalchemy-mate==1.4.28.4
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.38
stack-data==0.6.3
starlette==0.46.2
supervisor==4.2.5
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tf-playwright-stealth==1.1.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tldextract==5.1.3
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
TorchCRF==1.1.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
trendspy==0.1.6
trio==0.29.0
trio-websocket==0.12.2
triton==3.2.0
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
uszipcode==1.0.1
uvicorn==0.34.2
w3lib==2.3.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
yarg==0.1.9
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
